+++ src/weapon.h
@@ -3,8 +3,7 @@
 
 #include "player.h"
 
-#define N_WEAPONS               11
-#define N_ILLEGAL_2P_WEPS       1
+#define N_WEAPONS               10
 #define WEP_SPACEMOD            42
 #define WEP_SMALL_INIT_OFFSET   4    /* Offset between score and icon */
 #define WEP_SMALL_PADDING       25   /* Padding between small weapon icons */
@@ -33,8 +32,7 @@ enum weapons {
     WEP_WARP,
     WEP_GHOST,
     WEP_TRON,
-    WEP_CHILI_RUN,
-    WEP_SWITCH
+    WEP_CHILI_RUN
 };
 
 struct weapon {
@@ -55,7 +53,6 @@ int wepWarp(struct player *p, bool on);
 int wepGhost(struct player *p, bool on);
 int wepTron(struct player *p, bool on);
 int wepChilirun(struct player *p, bool on);
-int wepSwitch(struct player *p, bool on);
 
 static struct weapon wep_list[N_WEAPONS] = {
     {wepLightningspeed, 1, "Lightning-speed", "Gain lightning-speed.", ""},
@@ -67,8 +64,7 @@ static struct weapon wep_list[N_WEAPONS] = {
     {wepWarp, 4, "Warp", "Warp to a random spot", "on the map."},
     {wepGhost, 1, "Ghost walk", "Transform into ghost", "form."},
     {wepTron, 1, "Tron-mode", "No more smooth turns.", ""},
-    {wepChilirun, 1, "Chili run", "Hot hot hot!", ""},
-    {wepSwitch, 1, "Switch-aroo", "SWITCH-AROOOO!", ""}
+    {wepChilirun, 1, "Chili run", "Hot hot hot!", ""}
 };
 
 #endif